# ios/SplitBackAPI/Sources/SplitBackAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SplitBack",
    "version": "0.1.0"
  },
  "paths": {
...
    "/expenses/{expense_id}/receipts": {
      "post": {
        "tags": [
          "receipts"
        ],
        "summary": "Upload Receipt",
        "operationId": "upload_receipt_expenses__expense_id__receipts_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        }
      },
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "List Receipts",
        "operationId": "list_receipts_expenses__expense_id__receipts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "expense_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Expense Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReceiptResponse"
                  },
                  "title": "Response List Receipts Expenses  Expense Id  Receipts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}/content": {
      "get": {
        "tags": [
          "receipts"
        ],
        "summary": "Download Receipt",
        "operationId": "download_receipt_receipts__receipt_id__content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/receipts/{receipt_id}": {
      "delete": {
        "tags": [
          "receipts"
        ],
        "summary": "Delete Receipt",
        "operationId": "delete_receipt_receipts__receipt_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plaid/link-token": {
      "post": {
        "tags": [
          "plaid"
        ],
        "summary": "Create Link Token",
        "operationId": "create_link_token_plaid_link_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plaid/exchange": {
      "post": {
        "tags": [
          "plaid"
        ],
        "summary": "Exchange",
        "operationId": "exchange_plaid_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plaid/sync": {
      "post": {
        "tags": [
          "plaid"
        ],
        "summary": "Run Sync",
        "operationId": "run_sync_plaid_sync_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plaid/items": {
      "get": {
        "tags": [
          "plaid"
        ],
        "summary": "List Items",
        "operationId": "list_items_plaid_items_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaidItemResponse"
                  },
                  "type": "array",
                  "title": "Response List Items Plaid Items Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/plaid/items/{item_id}": {
      "delete": {
        "tags": [
          "plaid"
        ],
        "summary": "Delete Item",
        "operationId": "delete_item_plaid_items__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "plaid_account_id": {
            "type": "string",
            "title": "Plaid Account Id"
          },
          "balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "balance",
          "currency",
          "created_at",
          "updated_at"
        ],
        "title": "AccountResponse"
      },
...
      "ExchangeRequest": {
        "properties": {
          "public_token": {
            "type": "string",
            "title": "Public Token"
          },
          "user_identifier": {
            "type": "string",
            "title": "User Identifier",
            "default": "matt"
          },
          "institution_name": {
            "type": "string",
            "title": "Institution Name"
          }
        },
        "type": "object",
        "required": [
          "public_token"
        ],
        "title": "ExchangeRequest"
      },
      "ExchangeResponse": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          },
          "plaid_item_id": {
            "type": "string",
            "title": "Plaid Item Id"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            },
            "type": "array",
            "title": "Accounts"
          }
        },
        "type": "object",
        "required": [
          "item_id",
          "plaid_item_id",
          "accounts"
        ],
        "title": "ExchangeResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkTokenRequest": {
        "properties": {
          "user_identifier": {
            "type": "string",
            "title": "User Identifier",
            "default": "matt"
          }
        },
        "type": "object",
        "title": "LinkTokenRequest"
      },
      "LinkTokenResponse": {
        "properties": {
          "link_token": {
            "type": "string",
            "title": "Link Token"
          }
        },
        "type": "object",
        "required": [
          "link_token"
        ],
        "title": "LinkTokenResponse"
      },
...
      "PlaidItemResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plaid_item_id": {
            "type": "string",
            "title": "Plaid Item Id"
          },
          "institution_name": {
            "type": "string",
            "title": "Institution Name"
          },
          "user_identifier": {
            "type": "string",
            "title": "User Identifier"
          },
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/AccountResponse"
            },
            "type": "array",
            "title": "Accounts",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plaid_item_id",
          "created_at",
          "updated_at"
        ],
        "title": "PlaidItemResponse"
      },
      "ReceiptResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "expense_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expense Id"
          },
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "expense_id",
          "bucket",
          "object_key",
          "created_at"
        ],
        "title": "ReceiptResponse"
      },
...
      "SyncRequest": {
        "properties": {
          "item_id": {
            "type": "string",
            "format": "uuid",
            "title": "Item Id"
          }
        },
        "type": "object",
        "title": "SyncRequest"
      },
      "SyncResponse": {
        "properties": {
          "items_synced": {
            "type": "integer",
            "title": "Items Synced"
          },
          "accounts": {
            "type": "integer",
            "title": "Accounts"
          },
          "added": {
            "type": "integer",
            "title": "Added"
          },
          "modified": {
            "type": "integer",
            "title": "Modified"
          },
          "removed": {
            "type": "integer",
            "title": "Removed"
          }
        },
        "type": "object",
        "required": [
          "items_synced",
          "accounts",
          "added",
          "modified",
          "removed"
        ],
        "title": "SyncResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}